Port to Ubuntu. Mostly changes to CMakeFiles.txt and fixing compares of unsigned to signed warnings

+++ src/apps/ethslurp/ethslurp.cpp
@@ -172,7 +172,7 @@ bool CSlurperApp::Slurp(COptions& options, SFString& message) {
 
     // If the user tells us he/she wants to update the cache, or the cache
     // hasn't been updated in five minutes, then update it
-    uint32_t nSeconds = max(60ULL, toml.getConfigInt("settings", "update_freq", 300));
+    uint32_t nSeconds = max(60UL, toml.getConfigInt("settings", "update_freq", 300));
     if ((now - fileTime) > nSeconds) {
         // This is how many records we currently have
         uint32_t origCount  = theAccount.transactions.getCount();
@@ -309,7 +309,7 @@ bool CSlurperApp::Filter(COptions& options, SFString& message) {
         funcFilts[nFuncFilts++] = nextTokenClear(filtList, ',');
 
     theAccount.nVisible = 0;
-    for (int i = 0 ; i < theAccount.transactions.getCount() ; i++) {
+    for (uint32_t i = 0 ; i < theAccount.transactions.getCount() ; i++) {
         CTransaction *trans = &theAccount.transactions[i];
         trans->pParent = &theAccount;
 
@@ -343,7 +343,7 @@ bool CSlurperApp::Filter(COptions& options, SFString& message) {
 
         if (!options.funcFilter.empty()) {
             bool show = false;
-            for (int i = 0 ; i < nFuncFilts ; i++)
+            for (uint32_t i = 0 ; i < nFuncFilts ; i++)
                 show = (show || trans->isFunction(funcFilts[i]));
             trans->m_showing = show;
         }
@@ -367,7 +367,7 @@ bool CSlurperApp::Filter(COptions& options, SFString& message) {
     if (!isTesting) {
         double stop = qbNow();
         double timeSpent = stop-start;
-        fprintf(stderr, "\tFilter passed %llu visible records of %u in %f seconds\n",
+        fprintf(stderr, "\tFilter passed %lu visible records of %u in %f seconds\n",
                     theAccount.nVisible, theAccount.transactions.getCount(), timeSpent);
         fflush(stderr);
     }
@@ -385,7 +385,7 @@ bool CSlurperApp::Display(COptions& options, SFString& message) {
     if (!isTesting) {
         double stop = qbNow();
         double timeSpent = stop-start;
-        fprintf(stderr, "\tExported %lld records in %f seconds             \n\n", theAccount.nVisible, timeSpent);
+        fprintf(stderr, "\tExported %ld records in %f seconds             \n\n", theAccount.nVisible, timeSpent);
         fflush(stderr);
     }
     return true;
@@ -502,14 +502,14 @@ void findBlockRange(const SFString& json, uint32_t& minBlock, uint32_t& maxBlock
 
     minBlock = 0;
     int64_t first = json.find(search);
-    if (first != NOPOS) {
+    if (first != (int64_t)NOPOS) {
         SFString str = json.substr(first+len);
         minBlock = toLong32u(str);
     }
 
     SFString end = json.substr(json.ReverseFind('{'));  // pull off the last transaction
     int64_t last = end.find(search);
-    if (last != NOPOS) {
+    if (last != (int64_t)NOPOS) {
         SFString str = end.substr(last+len);
         maxBlock = toLong32u(str);
     }
